pyproject.toml
[project]
name = "newsy"
version = "0.1.0"
description = "AI・テクノロジーニュースをラジオ番組に変換して音声を生成する"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "feedparser>=6.0.11",
    "trafilatura>=1.12.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.0",
    "openai>=1.0.0",
    "soundfile>=0.12.0",
    "numpy>=1.26.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
